Add a shareable architecture diagram
The README's Mermaid diagram is the right thing for the README: GitHub
renders it natively and a screen reader can read it. It is the wrong thing
for anywhere that only accepts an image — a talk, a post, an issue on
another project. This adds the same story as an SVG and a rendered PNG.

The diagram is drawn to argue the product's thesis rather than to label
boxes: the flow runs watch, plan, approve, execute, with the human gate as
the only accented element, and the return arrow shows evictions going back
through the API server, which is what makes "your PDBs are enforced by
Kubernetes, not by dencer" visibly true rather than merely claimed.

render-social-preview.mjs becomes render-svg.mjs and takes its input and
output as arguments, because a second single-purpose copy of the same forty
lines is how the CLI build ended up with two implementations that disagreed
about a filename. `make social-preview` reproduces the committed PNG
byte-identically through the generalised renderer, which is the evidence
this refactor changed nothing.
